libutil: allow NULL default in environment_from_env()
Problem: There may be cases where we want to set an environment variable
from the current environment with environment_from_env() only if the
variable is currently set, but environment_from_env() does not allow this.

Allow the default variable value to be set to NULL, in which case
environment_from_env() does not set a default value for the requested
variable, it instead leaves the variable unset.