testsuite: add FLUX_SHELL_PLUGIN_NAME to shell plugins
Problem: A preprocessor definition of FLUX_SHELL_PLUGIN_NAME will
soon be required for shell plugins, in order for the shell_log()
and related macros to work, but this is not defined in the testsuite
shell plugins.

Add FLUX_SHELL_PLUGIN_NAME definition to all shell plugins in the
testsuite.