resource: fail get-xml request on quorum subset
Problem: if R is missing augmented graph information, fluxion
might block forever on resource.get-xml request for resource
discovery results.

Resource discovery is the natural fallback if resources not
not configured.  But fluxion cannot tell the difference between
R that lacks configured JGF augmentation because it was misconfigured,
and R allocated by a scheduler that doesn't support it.

If discovered hwloc results are not immediately available at the time
of the request, fail the request if the broker.quorum attribute is
less than the full instance, as this indicates that the instance
should begin operation with some ranks offline, and R should contain
all needed information.

Tack a test onto t3301-system-latestart.t

Fixes #3884.