job-manager: alloc-bypass: skip free if post of alloc event fails
Problem: If the post of the alloc event fails in the alloc-bypass
plugin, the "free" flag is still set on the job, which is not only
unnecessary, but may cause a spurious free event in the eventlog
during cleanup.

Skip addition of the "free" flag on jobs if posting of alloc event
fails.