shell: rc: load flux.shell module before sourcing initrc
Problem: There is no builtin way to extend the shell's Lua interfaces
in Lua for initrc scripts. Instead the shell currently uses a kludgy
00-lib.lua file in lua.d, which isn't guaranteed to be loaded if,
for example, the default initrc.lua is replaced.

Automatically load a "flux.shell" Lua module before running the
default initrc script. Core provided Lua-only functions can then be
placed into this module for a more reproducible environment whether
or not the default shell initrc is replaced or modified.

Log a debug message, but do not fail with an error, if the flux.shell
module fails to load.