shell: move spectrum mpi plugin to lua.d/mpi/spectrum.lua
Problem: The spectrum MPI plugin shouldn't be loaded unless the -o
mpi=spectrum option is used, but it sits in the main shell rcpath so
it is always loaded.

Move the spectrum MPI plugin to lua.d/mpi/spectrum.lua so that it is
only loaded with `-o mpi=spectrum`.