testsuite: add simulated broker failure tests
Problem: Tests are needed to ensure that when a broker dies, a job
can be found and monitored when the broker comes back up.  However, this
is extremely hard to do with the current broker.

Solution: In order to test libsdprocess and systemd when a broker
goes down, we simply simulate a broker executing a job and coming
back online through a simple sim-broker executable.