shell: pmi: do not use shell_die() in PMI_Abort()
Problem: When an MPI application calls MPI_Abort() and triggers
the PMI Abort handler in the shell's pmi plugin, shell_die() is used
to log the abort and raise a fatal job exception. However, shell_die()
exits immediately, and so job stderr with more error context could
be lost.

Log a job exception and return the shell's reactor in the PMI_Abort
handler, instead of using shell_die(). This should give the job shell
a chance to log pending stdout/err before the job-exec system terminates
the job.

Fixes #3894