docker: fix permissions on config files in fluxorama image
Problem: Possibly depending on current umask value, the security
significant toml config files in the fluxorama image are created
with 0664 permissions instead of 0644. This causes the IMP in the
resulting container to refuse to run jobs due to insecure permissions
on these files.

Explicitly set the permissions of these files to avoid potential
errors.