broker/overlay: drop overlay.health wait parameter
Problem: overlay.health "wait" parameter is not generally useful.

When the overlay.health RPC wait parameter is present, overlay.health
delays its response until a target subtree health state is reached.
This is only useful in contrived test cases.  To make the RPC more
generally useful, have it optionally stream updates each time the
state of the subtree or its children change.

Change the flux-overlay command to implement the --wait option in
the client, compensating for removal of optional wait parameter
in the server.  This way the tests don't need to be changed.

Drop one test of the RPC itself that no longer applies.