testsuite: fix t0016-cron-faketime.t on centos 8
Problem: The t0016-cron-faketime.t test is skipped even when
libfaketime is installed on centos8 because the libfaketime DSO is
installed in /usr/lib64/libfaketime/. Also, on centos8 results from
dumping cron entries are sometimes formatted as floating point,
but the tests expect integer.

Handle the centos8 location of libfaketime specifically in the test,
as was already done for the unique location on Ubuntu.

Format all "next_wakeup" values in the test as integer to avoid
breaking checks on centos8.