ci: fix undetected errors in checks_run.sh
Problem: The ci checks script, checks_run.sh, doesn't exit on some
fatal errors because of the use of `COMMAND || (printf "error" &&
exit 1)`, which only exits the subshell, not the script. This causes
the script to continue and get unrelated errors, which is confusing.

Introduce a new convenience function `checks_die()` which takes
an error message and some optional commands to run and use this
instead of a subshell.