testsuite: run_timeout.py: print an error on timeout
Problem: Sometimes it is difficult to determine that a test run
under t/scripts/run_timeout.py timed out when debugging test failures.
The script kills the process when it times out, but otherwise doesn't
log an error, and therefore this appears as silent failure.

Add a message to run_timeout.py when a process times out so there is
a hint in the test logs.