testsuite: fix t/scripts/stats-listen.py
Problem: The test stats receiver, t/scripts/stats-listen.py, doesn't
fail if the `-s, --search-for` option is used due to an indentation
error (the count of packets to wait for is only checked in the else:
clause). Also the script does not have any diagnostic output, making
failures difficult to diagnose.

Fix the indentation error so that the test will fail if the number
of packets received is less than the expected amount.

Add several diagnostic messages to aid in any future debugging.