docker: small updates for fluxorama Dockerfile
Problem: The fluxorama dockerfile has no way to start from a different
image file, and also unnecessarily installs psmisc and tmux in the
"systest" stage of the build.

Allow IMAGE (the base image for fluxorama) to be set as a build-arg,
and move the dnf install of psmisc,tmux after the systest stage.