broker: add tbon.prefertcp attribute
Problem: there is not a simple way to force the use of tcp://
when testing with flux start --test-size=N.

The PMI bootstrap code looks at broker.mapping to determine if all
the brokers are co-located and selects the ipc transport in that
case, otherwise tcp.  Add a check for the attribute 'tbon.prefertcp'.
If set to a non-zero integer value, then select tcp over ipc.

Example:
  flux start -s128 -o,-Stbon.prefertcp=1

Add a test to t0001-basic.t