broker: add tbon.zmqdebug attribute
Problem: zmq socket events are logged if a late-model zmq_monitor()
is available, but the logging is extremely verbose at times and not
always useful.

Disable zmq socket event logging by default, and allow it to be enabled
by setting the tbon.zmqdebug broker attribute to a nonzero value on the
command line.  Since it now must be actively enabled, log a non-fatal
error if setup fails.

Set tbon.zmqdebug=1 for the 'system' personality tests, and add a
test to t0001-basic.t.