python: do not auto-stop ProgressBar by default
Problem: Currently a ProgressBar instance calls self.stop()
automatically whenever self.count == self.total. However, this
resets the terminal when the calling program might not be ready
for that, which results in very messy output (For example the
program may still be issuing output which gets mixed up with
the final progressbar).

Do not call self.stop() by default for ProgressBar objects, instead
let the caller do it themselves when the terminal is ready to be
reset. Stopping the progress bar is not even compulsory as it will
be done in the atexit handler anyway.

Add an autostop attribute to get the old behavior if desired.