broker/groups: new class
Problem: the pattern of tracking broker rank membership
in a group using a timed batch reduction is repeated in
the broker state machine and resource module.

Implement a new "groups" class that can perform this
service using one code base for multiple sets.

In addition to implementing the timed batch reduction of
idset join/leave operations, register a callback with the
overlay so that when a TBON child goes down, the ranks
in the child's subtree are automatically removed from
any groups they are in.

Add t/scripts/groups.py test program.
Add t0027-broker-groups.t sharness test.