job-manager: support job manager prolog and epilog
Problem: There are use cases where a jobtap plugin needs to block or
pause a job after it has been allocated resources (after the `alloc`
event), but before the job starts running (before the `start` request
is made to the execution system), but currently the start request is
sent immediately after the alloc event.  Similarly, it would be useful
to block a job after the `finish` event, but before resources are
returned to the scheduler (before the `free` request to the scheduler).

Introduce the concept of job manager prolog and epilog actions.
Together, call these actions "perilogs" and add a `perilog_active`
count for every job. If the perilog_active count is nonzero, then
defer sending of start/free requests until the count is zero.

The perilog_active count is managed by RFC 21 prolog-{start,finish}
and epilog-{start,finish} events.