job-manager: send reattach flag to job-exec
Problem: When the job-manager is re-loaded discovers a job that's
in the RUN state, there is no way to inform the job exec module
that the job should still be running.

Solution: Add a "reattach" flag to the job-exec.start RPC.  This
flag informs the job-exec module that the job should still be
running.