job-manager: fix potential use-after-free of struct job
Problem:  If an event is posted for a job in CLEANUP state after
an exception in SCHED state (e.g. from a jobtap plugin in the
job.state.cleanup callback), then the job will be freed in a recursive
call to event_job_action(), resulting in use-after-free for that job
before the final call to event_job_action().

e.g. If a job is canceled while it is in SCHED state, the following
call trace occurs:

 event_job_post_entry ("exception") job->state=SCHED
  event_job_update() (old_state=SCHED, job->state=CLEANUP)
  event_jobtap_call("job.state.cleanup")
    -->post new event "test-event")
    event_job_post_entry() ("test-event") job->state=CLEANUP
       event_job_update() (old_state=CLEANUP job->state=CLEANUP)
       event_job_action(event="test-event")
         event_job_post_entry ("clean") job->state=CLEANUP
           event_job_update() (old_state=CLEANUP, job->state=INACTIVE)
           event_jobtap_call("job.state.inactive")
           event_job_action()
              --> remove job from active jobs hash, destroy job
  event_job_action() (for event "exception" but job is now destroyed)

A long term solution is post events to a queue and process them in
order in the job-manager via a reactor watcher, so that all events are
fully processed in the order they are submitted.

However, for now fix only the use-after-free bug by taking a reference
on the current job around the calls to event_jobtap_call() and
event_job_action().

Fixes #3919