doc: pin sphinx version to 3.4.3
Problem: sphinx versions <2 are incompatible with docutils >=0.18
and this breaks the automated readthedocs build. Projects created
before Oct 2020 use sphinx 1.8 by default, and encounter this bug
unless docutils is pinned to <0.18 or a newer sphinx version is
requested.

Update the sphinx version and specify an exact version as suggested
here:

  https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html#pinning-dependencies

to avoid future builds breaking unexpectedly.

Also pin docutils to <0.18 since it seems to cause issues with
sphinx<4.