broker/state_machine: use groups for quorum
Problem: state machine implements its own timed reduction
for quorum.

Convert the state machine to use the groups class.  Maintain
the set of brokers that have completed rc1 in a "broker.online"
group that can be made available for other uses.

A streaming groups.get is initiated on rank 0 for broker.online.
Upon entering the QUORUM state, each broker sends a groups.join
request.  When, in the groups.get continuation on rank 0,
broker.quorum is a subset of broker.online, the quorum-full event
is posted.  Brokers remain in the broker.online group until it
disconnects or is lost.