job-exec/testexec: emulate job reattachment
Problem: The testexec exec implementation does not parse and handle
the "reattach" flag from the job-manager.

Solution: If the testexec implementation sees the "reattach" flag from
the job-manager, emulate that job is still running by running the job
for the remaining time it should given the job's start time.  Emit
a "re-starting" event indicating this restart and notify the
job-manager via a "reattach" event.