flux-broker-attributes(7): update rundir desc
Problem: the description of the broker rundir attribute
gives the impression that it is a useful tunable for users,
but it is actually somewhat perilous given the UNIX socket
path length limit.

As proposed in #3925, update the description to warn about
the path limit and discourage casual use by users.