cmd: add flux-perilog-run.py
Problem: Flux currently only has a facility for running a single
job-manager prolog/epilog for jobs, but no easy way to run per-node
prolog or epilog commands before jobs start executing.

Add a helper script that is meant to fill this gap. The new command,
`flux-perilog-run`, can be used to execute prolog/epilog scripts locally
and/or across the nodes of a job, automatically draining nodes on
which commands fail.

The script can additionally run a set of scripts in a directory, by
default /etc/flux/system/prolog.d and /etc/flux/system/epilog.d.

This script is meant to be configured for the `perilog.so` jobtap
plugin. For example, the following will cause `flux-imp run prolog`
to be executed across a job's assigned ranks as part of the job-manager
prolog action.

[job-manager.prolog]
command = [ "flux", "perilog-run", "prolog",
            "-e/usr/libexec/flux/flux-imp,run,prolog" ]