flux-mini: print any --cc args if used in bulksubmit --dry-run
Problem: flux mini bulksubmit --dry-run doesn't display the effect of
--cc option, which is applied in a second pass of commandline expansion.
Instead the --dry-run output completely drops the --cc option, which
may bewilder users, e.g.

 $ flux mini bulksubmit --dry-run --cc=1-2 echo {} ::: 1 2 3
 flux-mini: submit echo 1
 flux-mini: submit echo 2
 flux-mini: submit echo 3

Since the --cc option is expanded in a second pass, it will be
difficult to get the final commandline in the --dry-run output. However,
a compromise is to include the supplied --cc option in the generated
dry-run output, so at least the option is not dropped, and users can
get an idea of what would be submitted, e.g.:

 $ flux mini bulksubmit --dry-run --cc=1-2 echo {} ::: 1 2 3
 flux-mini: submit --cc=1-2 echo 1
 flux-mini: submit --cc=1-2 echo 2
 flux-mini: submit --cc=1-2 echo 3

Fixes #3955