etc/rc1: Support FLUX_INSTANCE_RESTART
Problem: rc1 cancels all jobs when an instance is exited, but that may
not be desireable all of the time, such as some testing scenarios.

Solution: Support an environment variable FLUX_INSTANCE_RESTART to notify
rc1 that we instance restart is occurring and to not cancel jobs upon
instance shutdown.