resource: improve topo-reduce error handling
Problem: resource module logs "resource.topo-reduce: Cannot
allocate memory" if a duplicate request is received before
reduction is complete.

The resource.topo-reduce message handler silently ignores requests
after the reduction at that point in the TBON is complete, however
that is insufficent to cover the case of an early node restart.
If that occurs, rlist_append() fails.  Any rlist_append() failure is
assumed to be an ENOMEM, which stops the resource module reactor loop.

Change the logic to silently ignore the request if rlist_append()
fails, and don't stop the reactor on any other error in that function.

Fixes #3946