testsuite: cover duplicate topo-reduce messages
Problem: there is no test coverage for the resource
module receiving a duplicate topo-reduce message before
reduction is complete.

Fix the t2310-resource-module.t size at 4, then step through
module loading so that one of the nodes reloads the module
before all ranks of been loaded.  Prior to a recent fix
this was a fatal error to the resource module as described
in #3946.