libflux: save target nodeid of RPC in struct flux_rpc
Problem: It would be helpful during problem diagnosis if Flux
components log the target rank of RPCs during error handling, as this
can help pinpoint bad nodes and/or ranks after the fact. However, the
target rank of an RPC is not readily available from the RPC future,
which makes saving the rank an unnecessary burden on the component
which would like to issue the more useful error.

Add a nodeid member to `struct flux_rpc` and add an accessor.
This allows existing error handling functions to easily be extended
to add the failing target rank, without onerous boilerplate.

Note: the function returns the rank directly instead of returning by
reference as in `flux_msg_get_nodeid()` to allow direct use from Flux
logging functions.

If `flux_rpc_get_nodeid()` is called on a future that is not an RPC
then FLUX_NODEID_ANY is returned as a "catchall" value.