testsuite: allow t4000-issues-test-driver.t to run on a subset
Problem: When developing a new test for t4000-issues-test-driver.t,
it is tedious to wait for all existing tests to run before the new
test is executed by the test driver.

Add an option to adjust the default glob of "*" to something custom
which will match a subset of the tests in the t/issues directory, e.g.

 T4000_ISSUES_GLOB=t3920* ./t4000-issues-test-driver.t

would only run the test for issue 3920.