job-exec: improve job exception note on exec error
Problem: The note for job exceptions raised by the job-exec module
due to remote subprocess exceution errors is terse and probably
confusing to users. For example, when a remote rank goes offline a
direct translation of the EHOSTUNREACH error using strerror() will
result in the incomprehensible error "No route to host".

Customize the note in the EHOSTUNREACH error case, indicating that
contact to the job shell on the offending rank was lost.  For other
errors, expand the error to say "job shell execution error", which
may be more helpful to the user, since the exception will likely be
emitted to their job output logs.