broker: set hostlist attr in PMI bootstrap
Problem: hostlist attr is only set when the instance boostraps
from config files.

Add code to build the hostlist on each rank for the PMI case by
reading all the node business cards.  Our PMI performs a global
exchange on each fence.  The main overhead here is serialized fetching
of KVS entries for each rank.

This is performed in parallel across all the nodes rather than having
one rank do it and place the hostlist in the PMI KVS for others to fetch,
since the latter approach might overflow the maximum PMI KVS value size
if the hostlist is long (e.g. "flux start --test-size=N" where the
repeated hostname cannot be range-compressed).

Drop one test that verifies the hostlist is missing if not configured.