broker: rename config.hostlist attr to hostlist
Problem: we need to set hostlist generally, not just in the
config file bootstrap case, so the name "config.hostlist" is
not a great choice.

Rename the attribute to "hostlist".

Drop resource module assumption that a configured R should be re-ranked
if the hostlist attr is set, since that may no longer indicate that a
rank mapping was set in the [bootstrap] config.  Instead, check that
[bootstrap] exists (without parsing) before re-ranking.

Update tests and users.