broker: cache immutable attrs in broker flux_t
Problem: the broker adds rank and size to the flux_t
broker attribute cache so that flux_get_size() and flux_get_rank()
can function without making RPCs, but other attributes such as
hostlist would be useful to access from within the broker too.

Replace the one-off code for caching those two attributes with
a snapshot of all immutable (cacheable) attributes that are
available in the broker attribute hash when the function is called.