libflux: add flux_get_hostbyrank()
Problem: a fair amount of boilerplate is needed to use the
hostlist attribute to map a broker rank to its hostname.

Add flux_get_hostbyrank() which decodes the hostlist attribute and
stashes the result with the attribute cache in the flux_t handle.
The function always returns a printable value that does not need to
be freed so it can be used like this:

flux_log (h, LOG_ERR, "lost broker %s", flux_get_hostbyrank (rank));

Add unit test.