python: don't allow progressbar fraction above 1.0
Problem: If a programming mistake allows the fraction of filled
progressbar to go above 1.0, then the progressbar will begin to
wrap across multiple lines, potentially messing up the user's
terminal.

Do not allow the filled fraction to exceed 1.0 to avoid this effect.
The printed percentage, if enabled, will still show >100% however,
in case the overflow of total is intentional.