python: add a stop method to Futures objects
Problem: There is no convenient way to destroy a Python Future and
thus the underlying flux_future_t. This is inconvenient for
multi-response RPCs where we may want to disable any more callbacks
(after using cancel if appropriate).

Add a future.stop() method which simply disables any subsequent
callbacks to the Python continuation for the future. Since a
multi-response future is thus not reset, this effectively disables
the future, allowing the reactor to exit for instance.