python: add 'stop' parameter JobEventWatchFuture.cancel
Problem: When a Python then callback for a JobEventWatchFuture cancels
the event watch, it would be nice if the caller could (optionally) stop
future callbacks for events already queued from being invoked. Normally
this would be done from a C callback by destroying the future in
question, but this is not practical from Python.

Add an optional `stop` parameter (default False) to the
JobEventWatchFuture.cancel() which, when True, will call the stop()
method on the underlying future, disabling subsequent callbacks.