flux-top: add new command
Problem: there is no tool to provide "live view" of what is
happening in a Flux instance.

Add flux-top.

If run without arguments, it monitors the default Flux
instance.  If run with a job ID argument, and the job is found
to be a Flux instance (such as a batch job), that instance is
monitored via its remote-uri.

The top summary pane features a live view of the resource counts
provided by the "sched.resource-status" RPC.  In addition, it
contains:
- The jobid (if flux is running as a job in another flux instance)
- The instance depth
- The time left, if the instance was run with a time limit
- A blinking heart on each published heartbeat message
- The total number of pending, running, and inactive jobs.

The bottom job list pane provides a live view of jobs from the
"job-list.list" RPC.  The query includes filters so the response
only includes running jobs owned by the current user, and is
limited by the number of jobs that can be displayed in the pane.

The job runtimes and the instance expiration time need to be
recalculated periodically.  This is driven by the heartbeat event.

The resource summary, job statistics, and list of jobs are refreshed
as follows:  when a job-state event is received, a timer is started.
Further job-state events while the timer is running are ignored.
When the timer expires, async RPCs are initiated.  Data is refreshed
as the responses are received.  This rate-limits these heavy weight
queries to the timer interval, but ensures any updates are captured,
subject to the timer latency, while avoiding unnecessary queries when
nothing is happening.

Query results and time recalculations update the virtual screen.
The virtual screen is synced to the physical screen using a prepare
watcher that runs at the end of the event loop before poll(2).

A simple username cache is added to avoid looking up the same users on
every update.  The cache grows for the life of the flux-top invocation.

Some caveats for this first cut:
- The window geometry is assumed to be 80 colums wide.
  Field spacing is not adjusted based on the actual window width,
  nor does resizing the window readjust the field spacing.
- Shrinking a window and growing it again will not immediately restore
  obscured text, though it will fix itself on the next update.
  If all else fails, the screen can be redrawn by typing ctrl-l.
- Jobs are displayed in the order returned by the query.
- Displayed columns are not configurable.
- There is no help screen.