python: fix early garbage collection of in-use Watchers
Problem: If Python code creates an instance of one of the Watcher
subclasses (TimerWatcher, FDWatcher, SignalWatcher, etc), but does
not save a reference to the object, then the Python Watcher may
be garbage collected before the underlying C object is destroyed.
This can cause problems like a timer suddenly failing to fire, or a
crash when the C watcher calls back into Python and cannot associate
the C watcher with any Python watcher.

Stash all created Watcher objects in the Flux handle that was used
to create them, so that they are not eligible for GC until the Flux
handle is or `Watcher.destroy()` is manually called.

Fixes #3973