job-list: reduce code duplication in events handling
Problem: There is some duplicated code in job-list when handling
journal events. For instance, each specific event handler separately
looks up the job by id in the job hash and then advances the sequence
number (except for one, the annotations handler).

Always attempt to lookup the job before processing the current event.
If the job is not found and the event was not a submit event, then
log an error and return (as _lookup_job() used to do). Add special
handling for event sequence numbers < 0 to job_update_eventlog_seq()
so that sequence advancement can also be handled in a single location.

This moves most of the duplicated code to a single location.