job-manager: remove EVENT_JOURNAL_ONLY flag
Problem: There will be a need to support ephemeral events in the
job manager which are not posted to the job eventlog in the KVS, but
otherwise are processed the same as other events. Since these events
are not posted to the eventlog, their sequence number should not be
updated, but currently EVENT_NO_COMMIT increments the job eventlog
sequence number. EVENT_JOURNAL_ONLY also can't be used since it only
posts the event to the journal and doesn't do other event processing.

Drop the EVENT_JOURNAL_ONLY flag, since it has overlap with
EVENT_NO_COMMIT and it is probably better to allow these events to
be processed in the job manager event machinery the same as all other
events (e.g. jobtap callbacks)

Do not update eventlog sequence number with EVENT_NO_COMMIT to avoid
the journal and KVS eventlog sequences getting out of sync as described
above. Provide a new flag EVENT_FORCE_SEQUENCE for the case where a
sequence number should still be used with EVENT_NO_COMMIT (as in the
submit event use case)

This change means that the "annotations" events, which were previously
hidden from jobtap plugins are now available. Update tests accordingly.