flux-job: add flux-job memo subcommand
Problem: A new command is necessary to post RFC 21 memo events
to the job eventlog via the job-manager.memo service.

Add `flux-job memo` for this purpose. This command takes a set
of key=value pairs on the commandline and builds a single memo
object to send along to the job-manager memo service. As in
`flux kvs put`, keys can take dotted paths to allow easy
construction of more complex objects  without undue quoting of
raw JSON on the commandline.

A --volatile option is provided to prevent the memo object from
being posted to the job eventlog (Thus the memo update will be
lost on Flux restart)