job-manager: add job-manager.memo service
Problem: Currently no service exists to allow an external entity to
post an RFC 21 memo event to a job.

Add a job-manager.memo service endpoint which posts a memo event for
the target job, given sufficient permission.

The request takes a jobid, the JSON object used to update the
job's memo object, and an optional "volatile" parameter which is
used to determine if the memo event is posted to the eventlog with
EVENT_NO_COMMIT or not.