python: fix message refcount management
Problem: Python does not corrently manage reference counts for
Message objects, which allows the underlying flux_msg_t to possibly
be destroyed while the Python Message object instance is still
active in Python. This can lead to unpredictable results, crashes,
and random memory corruption.

Always take a reference to the underlying flux_msg_t when instantiating
a Message object, and always remove the reference when the Message
is garbage collected. This should ensure that the C flux_msg_t is
always around while the Python object is.

Fixes #3982