broker: set a uri job memo in parent instance
Problem: Checking if a Flux job is a child instance and fetching the
associated URI is currently less convenient than it could be since
the URIs are stored in the job's private namespace KVS.

Now that job "memo" events are supported, have the rank 0 broker
post a memo for the job's URI if the "jobid" and "parent-uri" attributes
are present.

The end result is that the URI for child jobs in a Flux instance
is available directly in the `flux jobs` job listing via the
`user.uri` attribute.