libflux: avoid EINVAL on NULL return param
Problem: the API is inconsistent with respect to whether
a function fails with EINVAL when passed a NULL parameter
that is used to pass back a return value by reference.

Since the user hasn't provided storage for a result, there
is little chance that they are misusing the function
unintentionally.  Instead of returning EINVAL in this
case, return success and simply don't assign the value.

Update unit tests to check for segfault rather than check
for an EINVAL error.