libflux: avoid double close on fd in flux_close()
Problem: flux_close() calls flux_handle_destroy(), which closes the
file descriptor it obtained from connector->pollfd(), but that file
descriptor is owned by the implementation, and should not be closed
by users.

N.B. most likely this exists because in the shmem connector, the fd is
obtained by calling zsock_fd() on the zsock_t object, and zsock(3) says:

  Caller owns return value and must destroy it when done

However, zsock_fd() is just a wrapper for zmq_getsockopt (ZMQ_FD),
and zmq_getsockopt(3) says:

  CAUTION: The returned file descriptor is intended for use with a 'poll'
  or similar system call only. Applications must never attempt to read
  or write data to it directly, neither should they try to close it.

Most likely the zsock(3) documentation is incorrect.

In the other connector implementations, it is more obvious that the fd
belongs to the implementation and is closed by connnector->impl_destroy().

Remove extra close from flux_handle_destroy().