broker: register event service on all ranks
Problem: the broker "event" service is only registered on rank 0
with only the "pub" method, but we should consolidate broker publishing
and subscribing under one service and subscription will need to be
on all ranks.

Register the service on all ranks, and change client to always send
the RPC to rank 0 rather than let the broker route it upstream as an
unmatched service.  Fail event.pub if received on a rank > 0.  This is
not likely as it is only used internally.

The prototype for publisher_create() was modified to pass in the broker
context, for access to the rank.  This was not strictly necessary but
a commit following this one will take advantage of that to move the
subscribe code into this source module, and then it will be required.