broker: rename broker.sub to event.subscribe
Problem: broker has a unique service name for subscribe/unsubscribe,
but we want to unify event subscription RPCs so they work the same
in the broker and in "routers".

Rename broker.sub to event.subscribe
Rename broker.unsub to event.unsubscribe

Update the shmem connector to use the new name.