testsuite: add t9000-system.t
Problem: A mechanism for adding and running tests in the system test
CI builder does not exist.

Add a new sharness test t9000-system.t which runs a set of "system"
tests stored in t/system/*. The test runs using the *installed* version
of Flux, and is skipped unless FLUX_ENABLE_SYSTEM_TESTS is set in
the environment (e.g. by CI), or if the test is run with -d, --debug,
and a system instance appears to be present by virtue of a security.user
attribute set with a different userid than the current user.

The t9000-system.t test driver is set up slightly different from the issues
test driver. It sources each test file in `t/system/*` so that the test
files can individually make use of `test_expect_success()`.

The main reason for existence of the t9000-system.t sharness tests is
so that, at least for now, all system tests can be run serially so that
the individual system tests do not conflict in requests for resources
from the enclosing system instance.