python: add flux.uri.FluxURIResolver class
Problem: Flux does not have a common place to collect the methods
for discovery of URIs for Flux jobs. This lack causes an excess of
ad-hoc solutions across the Flux user-base and reduces the overall
user experience when working with hierarchical instances of Flux.

Introduce a new Python FluxURIResolver class which is meant to act as
an extensible repository of methods for discovery of URIs from other
sources. On initialization, the class discovers a set of URI resolver
"plugins" which can translate a simple URI given in the plugin's
"scheme" to a job URI. If a URI does not include a scheme, it is
assumed to be a Flux jobid (e.g.  "f1234" is resolved as "jobid:f1234")

As a convenience, a resolver URI may have an optional query component
of "remote" or "local" which will force the result into a local or
remote job URI. For intance

 jobid:f1234?local

would rewrite the discovered URI for job f1234 as a local:// URI.

Query parameters are _also_ passed into resolver plugins, and the use
of `local` or `remote` (or other plugin-supported query parameters)
may influence the plugin's URI resolution as well.