python: add jobid URI resolver plugin
Problem: There is no way to easily resolve a Flux instance running as
a job to its URI.

Add a "jobid" resolver plugin for the FluxURIResolver class.

This plugin resolves a Flux jobid or hierarchy of Flux jobids to a
URI by querying the user.uri job memo.

A hierarchy of jobs is specified by use of a forward-slash, e.g.

 "jobid:f1234/f3456"

would resolve the URI for job f3456 running in the job f1234.