python: add pid URI resolver plugin
Problem: It would useful to fetch the FLUX_URI from a local process,
but currently the schemes to do this are ad-hoc.

Add a "pid" resolver plugin to the FluxURIResolver class.

This plugin attempts to read the target process id's FLUX_URI from
/proc/pid/environ, allowing the user to connect to the same Flux
instance in which a process is running.

As a convenience, if the target PID is a running flux-broker, get
FLUX_URI from one of its children, allowing the "pid" resolver scheme
to fetch the FLUX_URI for a broker running on the current system.

This will be useful when discovering URIs when Flux is running under
a foreign resource manager.